frame: structural improvements
Package refactoring
- Debug flag is put to debug file
- Buffer writes are put to buffer_write.go file
- Buffer reads are put to buffer_read.go file
- errors are put to errors.go file

Read is refactored so that there are two functions readByte and readInto that are used by all the other functions and are responsible for handling the read error.
