+++ frame/debug.go
@@ -1,0 +1,3 @@
+package frame
+
+var Debug = true
